lab10_gio почти доправил фигуру. Сейчас даже красиво

+++ labs/lab10/mvvm/src/types/geometry/geometry.go
@@ -1,13 +1,13 @@
 package geometry
 
 type Vector struct {
-	X float64
-	Y float64
+	X int
+	Y int
 }
 
 type Point struct {
-	X float64
-	Y float64
+	X int
+	Y int
 }
 
 type Rect struct {
@@ -21,8 +21,8 @@ type Scale struct {
 }
 
 type Bounds struct {
-	Width  float64
-	Height float64
+	Width  int
+	Height int
 }
 
 func CalculateScale(
